requirements.txt
Django==5.1.4
django-rest-framework==0.1.0
googletrans==4.0.0rc1
joblib==1.3.2
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.0
pandas==2.1.1
pytorch-lightning==2.2.5
scikit-learn==1.3.1
seaborn==0.13.2
sentence-transformers==4.1.0
torch==2.7.0
wordcloud==1.9.3
regex==2023.8.8
